+++ apps/gizual-app/src/primitives/input/input.module.scss.d.ts
@@ -1,0 +1,4 @@
+declare const classNames: {
+  readonly Input: "Input";
+};
+export = classNames;