+++ apps/gizual-app/src/primitives/query-input/simple/modules/type/type-placeholder-module.tsx
@@ -26,7 +26,7 @@ export function TypePlaceholderModule() {
       isOpen={isOpen}
       setIsOpen={setIsOpen}
     >
-      <TypePlaceholderModal isOpen={isOpen} setIsOpen={setIsOpen} />
+      <TypePlaceholderModal closeModal={() => setIsOpen(false)} />
     </DialogProvider>
   );
 }