+++ apps/gizual-app/src/primitives/query-input/simple/simple-query-input.tsx
@@ -2,14 +2,24 @@ import { IconCommandLine } from "@app/assets";
 import { DialogProvider } from "@app/primitives/dialog-provider";
 import { IconButton } from "@app/primitives/icon-button";
 import { Tooltip } from "@mantine/core";
+import { notifications } from "@mantine/notifications";
 import { observer } from "mobx-react-lite";
+import React from "react";
 
+import { useQuery } from "@giz/maestro/react";
 import { AdvancedEditor } from "../advanced/advanced-query-input";
+import { QueryViewModel } from "../query.vm";
 
 import { ModuleProvider } from "./module-provider";
 import style from "./simple-query-input.module.scss";
 
 export const SimpleQueryInput = observer(() => {
+  // Since the Monaco component contains some of the validation logic,
+  // it's easier to share state via MobX than to try and pass props around.
+  const queryVm = React.useMemo(() => new QueryViewModel(), []);
+  const { setQuery } = useQuery();
+  const [isOpen, setIsOpen] = React.useState(false);
+
   return (
     <div className={style.Container}>
       <div className={style.Clickable}>
@@ -29,8 +39,35 @@ export const SimpleQueryInput = observer(() => {
               </Tooltip>
             }
             triggerClassName={style.AdvancedSearchIconTrigger}
+            contentClassName={style.AdvancedSearchDialog}
+            isOpen={isOpen}
+            setIsOpen={setIsOpen}
+            withFooter
+            defaultFooterOpts={{
+              cancelLabel: "Close",
+              okLabel: "Run Query",
+              hasOk: true,
+              hasCancel: true,
+              onOk: () => {
+                const validationResult = queryVm.validatedQuery;
+
+                if (validationResult.isValid) {
+                  setQuery(validationResult.query);
+                  setIsOpen(false);
+                } else {
+                  notifications.show({
+                    title: "Invalid Query",
+                    message: queryVm.validationOutput,
+                    color: "red",
+                  });
+                }
+              },
+              onCancel: () => {
+                setIsOpen(false);
+              },
+            }}
           >
-            <AdvancedEditor />
+            <AdvancedEditor vm={queryVm} />
           </DialogProvider>
         </div>
       </div>